refactor: 상품 데이터베이스 패키지 이동 및 상품 ID 추가
